requirements.txt
Django==3.2.16
Pillow==9.3.0
djangorestframework==3.14.0
djoser==2.1.0
drf-extra-fields==3.4.1
